+++ client/cody-shared/src/codebase-context/index.ts
@@ -17,7 +17,7 @@ export class CodebaseContext {
     constructor(
         private config: Pick<Configuration, 'useContext'>,
         private embeddings: EmbeddingsSearch | null,
-        private keywords: KeywordContextFetcher
+        private keywords: KeywordContextFetcher | null
     ) {}
 
     public onConfigurationChange(newConfig: typeof this.config): void {
@@ -26,17 +26,19 @@ export class CodebaseContext {
 
     public async getContextMessages(query: string, options: ContextSearchOptions): Promise<ContextMessage[]> {
         switch (this.config.useContext) {
-            case 'blended':
+            case 'embeddings' || 'blended':
                 return this.embeddings
                     ? this.getEmbeddingsContextMessages(query, options)
                     : this.getKeywordContextMessages(query, options)
-            case 'embeddings':
-                return this.getEmbeddingsContextMessages(query, options)
             case 'keyword':
                 return this.getKeywordContextMessages(query, options)
             default:
-                return []
+                return this.getEmbeddingsContextMessages(query, options)
         }
+    }
+
+    public checkEmbeddingsConnection(): boolean {
+        return !!this.embeddings
     }
 
     // We split the context into multiple messages instead of joining them into a single giant message.
@@ -79,6 +81,10 @@ export class CodebaseContext {
     }
 
     private async getKeywordContextMessages(query: string, options: ContextSearchOptions): Promise<ContextMessage[]> {
+        if (!this.keywords) {
+            return []
+        }
+
         const results = await this.keywords.getContext(query, options.numCodeResults + options.numTextResults)
         return results.flatMap(({ content, fileName }) => {
             const messageText = populateCodeContextTemplate(content, fileName)