+++ enterprise/dev/ci/internal/ci/bazel_operations.go
@@ -27,7 +27,7 @@ func BazelOperations() *operations.Set {
 	return ops
 }
 
-func bazelRawCmd(args ...string) string {
+func bazelCmd(args ...string) string {
 	pre := []string{
 		"bazel",
 		"--bazelrc=.bazelrc",
@@ -42,7 +42,7 @@ func bazelRawCmd(args ...string) string {
 // are correct, but do not actually build anything.
 func bazelAnalysisPhase() func(*bk.Pipeline) {
 	// We run :gazelle since 'configure' causes issues on CI, where it doesn't have the go path available
-	cmd := bazelRawCmd(
+	cmd := bazelCmd(
 		"build",
 		"--nobuild", // this is the key flag to enable this.
 		"//...",
@@ -51,7 +51,7 @@ func bazelAnalysisPhase() func(*bk.Pipeline) {
 	cmds := []bk.StepOpt{
 		bk.Key("bazel-analysis"),
 		bk.Agent("queue", "bazel"),
-		bk.RawCmd(cmd),
+		bk.Cmd(cmd),
 	}
 
 	return func(pipeline *bk.Pipeline) {
@@ -91,13 +91,12 @@ func bazelTest(targets ...string) func(*bk.Pipeline) {
 		"//client/web:bundlesize-report",
 	}
 
-	bazelTestCmd := bazelRawCmd(fmt.Sprintf("test %s", strings.Join(targets, " ")))
-	bazelRunCmd := bazelRawCmd(fmt.Sprintf("run %s", strings.Join(runTargets, " ")))
+	bazelTestCmd := bazelCmd(fmt.Sprintf("test %s", strings.Join(targets, " ")))
+	bazelRunCmd := bazelCmd(fmt.Sprintf("run %s", strings.Join(runTargets, " ")))
 	cmds = append(
 		cmds,
-		// TODO(JH): run server image for end-to-end tests on SOURCEGRAPH_BASE_URL similar to run-bazel-server.sh.
-		bk.RawCmd(bazelTestCmd),
-		bk.RawCmd(bazelRunCmd),
+		bk.Cmd(bazelTestCmd),
+		bk.Cmd(bazelRunCmd),
 	)
 
 	return func(pipeline *bk.Pipeline) {
@@ -113,13 +112,13 @@ func bazelBackCompatTest(targets ...string) func(*bk.Pipeline) {
 		bk.Agent("queue", "bazel"),
 
 		// Generate a patch that backports the migration from the new code into the old one.
-		bk.RawCmd("git diff origin/ci/backcompat-v5.0.0..HEAD -- migrations/ > dev/backcompat/patches/back_compat_migrations.patch"),
+		bk.Cmd("git diff origin/ci/backcompat-v5.0.0..HEAD -- migrations/ > dev/backcompat/patches/back_compat_migrations.patch"),
 	}
 
-	bazelRawCmd := bazelRawCmd(fmt.Sprintf("test %s", strings.Join(targets, " ")))
+	bazelRawCmd := bazelCmd(fmt.Sprintf("test %s", strings.Join(targets, " ")))
 	cmds = append(
 		cmds,
-		bk.RawCmd(bazelRawCmd),
+		bk.Cmd(bazelRawCmd),
 	)
 
 	return func(pipeline *bk.Pipeline) {
@@ -134,8 +133,8 @@ func bazelTestWithDepends(optional bool, dependsOn string, targets ...string) fu
 		bk.Agent("queue", "bazel"),
 	}
 
-	bazelRawCmd := bazelRawCmd(fmt.Sprintf("test %s", strings.Join(targets, " ")))
-	cmds = append(cmds, bk.RawCmd(bazelRawCmd))
+	bazelRawCmd := bazelCmd(fmt.Sprintf("test %s", strings.Join(targets, " ")))
+	cmds = append(cmds, bk.Cmd(bazelRawCmd))
 	cmds = append(cmds, bk.DependsOn(dependsOn))
 
 	return func(pipeline *bk.Pipeline) {
@@ -152,8 +151,8 @@ func bazelBuild(optional bool, targets ...string) func(*bk.Pipeline) {
 	cmds := []bk.StepOpt{
 		bk.Agent("queue", "bazel"),
 	}
-	bazelRawCmd := bazelRawCmd(fmt.Sprintf("build %s", strings.Join(targets, " ")))
-	cmds = append(cmds, bk.RawCmd(bazelRawCmd))
+	bazelRawCmd := bazelCmd(fmt.Sprintf("build %s", strings.Join(targets, " ")))
+	cmds = append(cmds, bk.Cmd(bazelRawCmd))
 
 	return func(pipeline *bk.Pipeline) {
 		if optional {